Rework the sift process to be more useful and efficient
Previously, `sift` applied post parse filters. However, parsing is
the most time consuming part of the job, and it's very easy to
parse the final stream on the fly, so it wasn't very useful.

The new filters are applied before parsing, on a `str` line level,
which prevents unneccessary parsing. I added four different filters
now, one of which is a `Callable` to allow custom filters.

This is not accurate enough to filter on e.g. `mobile == True`, but
that can easily be applied after calling `pvcreek`.